Injected move actions if existing move is invalid. Changed getRoom()
If the destination in the original move is 2 steps from the player's current location, inject a move to a common neighbor and modify original move to originate from the common nbr.
Unrelated: changed the way getRoom() works when a point is in multiple rooms. Now, such a point is assigned to the first of these multiple rooms in lex order. As long as a point is assigned the same room every time, we should in theory be ok.